_splat.c: add source helpers + tidy-up
Add `splat_check_all_floats' to test whether a series of
Python objects are floats using a variadic function.  This
can be used to optimise the code when only floats are used,
to avoid going through the signal vectors.

Also add some comments to separate different sections in
the file.